chore(*): move mocha tests to original test directory
NODE-1096
